# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Orpheus Space API
  version: 0.1.0
  description: Schedule management and stored history. Catalog and result endpoints read Orpheus synchronously; writes validate new profile and template selections.
  license:
    name: MIT
    url: https://opensource.org/license/mit
servers:
- url: http://localhost:8010
security:
- bearerAuth: []
- browserSession: []
- {}
...
paths:
  "/api/v1/schedules":
    get:
      operationId: ListSchedules
      summary: List schedules
      tags:
      - Schedules
      parameters:
      - "$ref": "#/components/parameters/OwnerEmail"
      - "$ref": "#/components/parameters/Unowned"
      - "$ref": "#/components/parameters/Status"
      - "$ref": "#/components/parameters/Limit"
      - "$ref": "#/components/parameters/Cursor"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SchedulePage"
        default:
          "$ref": "#/components/responses/Problem"
    post:
      operationId: CreateSchedule
      summary: Create a schedule
      description: SAML non-admins can create only for their own email; an omitted owner is filled from the session. Explicit null or another owner returns 403 schedule_forbidden. Bearer, anonymous and admins have full access.
      tags:
      - Schedules
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateSchedule"
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Schedule"
        default:
          "$ref": "#/components/responses/Problem"
  "/api/v1/schedules/{id}":
    parameters:
    - "$ref": "#/components/parameters/ID"
    get:
      operationId: GetSchedule
      summary: Get a schedule including deleted records
      tags:
      - Schedules
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Schedule"
        default:
          "$ref": "#/components/responses/Problem"
    patch:
      operationId: UpdateSchedule
      summary: Update selected schedule fields
      description: SAML non-admins can update only their own schedules and cannot change or clear the owner. Denied writes return 403 schedule_forbidden.
      tags:
      - Schedules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateSchedule"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Schedule"
        default:
          "$ref": "#/components/responses/Problem"
    delete:
      operationId: DeleteSchedule
      summary: Soft delete a schedule
      description: SAML non-admins can delete only their own schedules. Denied writes return 403 schedule_forbidden.
      tags:
      - Schedules
      responses:
        '204':
          description: Success
        default:
          "$ref": "#/components/responses/Problem"
  "/api/v1/schedules/profiles":
    get:
      operationId: GetProfiles
      summary: Get Orpheus profiles and the creation default
      tags: [Schedules]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Profiles"
        default:
          "$ref": "#/components/responses/Problem"
  "/api/v1/schedules/templates":
    get:
      operationId: GetTemplates
      summary: Get Orpheus templates and the creation default
      tags: [Schedules]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Templates"
        default:
          "$ref": "#/components/responses/Problem"
  "/api/v1/schedules/settings":
    get:
      operationId: GetSettings
      summary: Get allowed environment names
      tags:
      - Schedules
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Settings"
        default:
          "$ref": "#/components/responses/Problem"
  "/api/v1/schedules/preview":
    post:
      operationId: PreviewSchedule
      summary: Preview five future cron occurrences
      tags:
      - Schedules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PreviewInput"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Preview"
        default:
          "$ref": "#/components/responses/Problem"
...
  "/api/v1/schedules/{id}/occurrences":
    parameters:
    - "$ref": "#/components/parameters/ID"
    get:
      operationId: ListOccurrences
      summary: Read stored history without calling the core
      tags:
      - Schedules
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OccurrencePage"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - "$ref": "#/components/parameters/Limit"
      - "$ref": "#/components/parameters/Cursor"
  "/api/v1/schedules/{id}/occurrences/{occurrence_id}":
    parameters:
    - "$ref": "#/components/parameters/ID"
    - "$ref": "#/components/parameters/OccurrenceID"
    get:
      operationId: GetOccurrence
      summary: Read a stored occurrence without calling the core
      tags:
      - Schedules
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Occurrence"
        default:
          "$ref": "#/components/responses/Problem"
  "/api/v1/schedules/{id}/occurrences/{occurrence_id}/result":
    parameters:
    - "$ref": "#/components/parameters/ID"
    - "$ref": "#/components/parameters/OccurrenceID"
    get:
      operationId: GetOccurrenceResult
      summary: Explicitly fetch the current result from the Orpheus core
      tags:
      - Schedules
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OccurrenceResult"
        default:
          "$ref": "#/components/responses/Problem"
  "/api/v1/schedules/{id}/reset-session":
    parameters:
    - "$ref": "#/components/parameters/ID"
    post:
      operationId: ResetSession
      summary: Detach the reusable session for the next occurrence
      description: SAML non-admins can reset only their own schedules. Denied writes return 403 schedule_forbidden.
      tags:
      - Schedules
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Schedule"
        default:
          "$ref": "#/components/responses/Problem"
...
components:
  securitySchemes:
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-orpheus_space_session
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
        format: uuid
      description: Retries with the same normalized body return the original response.
        Keys do not expire.
    OwnerEmail:
      name: owner_email
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      description: Repeated email filter; OR semantics, at most 100 distinct normalized
        emails.
    Unowned:
      name: unowned
      in: query
      schema:
        type: boolean
      description: Only schedules without an owner. Cannot be combined with owner_email.
    Status:
      name: status
      in: query
      schema:
        "$ref": "#/components/schemas/Status"
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        minLength: 1
        maxLength: 8192
      description: Opaque position bound to normalized filters. New inserts are excluded
        from an ongoing traversal.
    OccurrenceID:
      name: occurrence_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: Structured API error. 401 credentials, 403 CSRF or schedule_forbidden, 404 missing, 409
        conflict, 422 validation, 503 unavailable.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
          - code
          - message
          - phase
          - details
          properties:
            code:
              type: string
            message:
              type: string
            phase:
              type: string
              nullable: true
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                required:
                - path
                - code
                properties:
                  path:
                    type: array
                    items:
                      type: string
                  code:
                    type: string
    Status:
      type: string
      enum:
      - active
      - paused
    SessionMode:
      type: string
      enum:
      - new
      - reuse
    EnvFrom:
      type: array
      uniqueItems: true
      items:
        type: string
        pattern: "^[A-Za-z_][A-Za-z0-9_]*$"
    Profile:
      type: object
      additionalProperties: false
      required: [name, description, harness, model, codex, instructions, is_default]
      properties:
        name: {type: string}
        description: {type: string, nullable: true}
        harness: {type: string, enum: [codex]}
        model: {type: string, nullable: true}
        codex: {$ref: '#/components/schemas/CodexProfile'}
        instructions: {type: string}
        is_default: {type: boolean}
    CodexProfile:
      type: object
      additionalProperties: false
      properties:
        effort: {type: string, enum: [none, minimal, low, medium, high, xhigh, max, ultra]}
        summary: {type: string, enum: [auto, concise, detailed, none]}
        personality: {type: string, enum: [none, friendly, pragmatic]}
        service_tier: {type: string, minLength: 1}
    Profiles:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Profile'}
    Template:
      type: object
      additionalProperties: false
      required: [name, description, is_default]
      properties:
        name: {type: string}
        description: {type: string, nullable: true}
        is_default: {type: boolean}
    Templates:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Template'}
    CreateSchedule:
      type: object
      additionalProperties: false
      required:
      - name
      - prompt
      - cron
      - timezone
      properties:
        profile:
          type: string
          minLength: 1
          description: Exact Orpheus profile name; creation uses the configured default when omitted.
        template:
          type: string
          minLength: 1
          description: Exact Orpheus template name; creation uses the configured default when omitted.
        name:
          type: string
          minLength: 1
          maxLength: 200
        prompt:
          type: string
          minLength: 1
        cron:
          type: string
          minLength: 1
        timezone:
          type: string
          minLength: 1
        status:
          "$ref": "#/components/schemas/Status"
        model:
          type: string
          nullable: true
          minLength: 1
        session_mode:
          "$ref": "#/components/schemas/SessionMode"
        owner_email:
          type: string
          nullable: true
          minLength: 1
          description: SAML non-admins must use their session email; omission fills it, explicit null is forbidden. Full-access callers may use any owner or null.
        env_from:
          "$ref": "#/components/schemas/EnvFrom"
    UpdateSchedule:
      type: object
      additionalProperties: false
      properties:
        profile:
          type: string
          minLength: 1
          description: Exact Orpheus profile name; omitted keeps the stored choice; a change starts a new reusable session.
        template:
          type: string
          minLength: 1
          description: Exact Orpheus template name; omitted keeps the stored choice; a change starts a new reusable session.
        name:
          type: string
          minLength: 1
          maxLength: 200
        prompt:
          type: string
          minLength: 1
        cron:
          type: string
          minLength: 1
        timezone:
          type: string
          minLength: 1
        status:
          "$ref": "#/components/schemas/Status"
        model:
          type: string
          nullable: true
          minLength: 1
        session_mode:
          "$ref": "#/components/schemas/SessionMode"
        owner_email:
          type: string
          nullable: true
          minLength: 1
        env_from:
          "$ref": "#/components/schemas/EnvFrom"
    Schedule:
      type: object
      additionalProperties: false
      required:
      - id
      - url
      - can_edit
      - name
      - prompt
      - cron
      - timezone
      - status
      - model
      - profile
      - template
      - session_mode
      - owner_email
      - env_from
      - created_at
      - updated_at
      - next_run_at
      - deleted_at
      - last_occurrence
      properties:
        can_edit:
          type: boolean
          description: Whether this caller may modify the current schedule. False for deleted schedules. Computed from current ownership even on an idempotent creation replay; busy state can still prevent session reset.
        profile:
          type: string
          minLength: 1
          description: Stored Orpheus profile name.
        template:
          type: string
          minLength: 1
          description: Stored Orpheus template name.
        url:
          type: string
          format: uri
          nullable: true
          description: Absolute Space Web card URL from ORPHEUS_PUBLIC_URL, or null when not configured. Computed at response time.
        name:
          type: string
          minLength: 1
          maxLength: 200
        prompt:
          type: string
          minLength: 1
        cron:
          type: string
          minLength: 1
        timezone:
          type: string
          minLength: 1
        status:
          "$ref": "#/components/schemas/Status"
        model:
          type: string
          nullable: true
          minLength: 1
        session_mode:
          "$ref": "#/components/schemas/SessionMode"
        owner_email:
          type: string
          nullable: true
          minLength: 1
        env_from:
          "$ref": "#/components/schemas/EnvFrom"
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        next_run_at:
          type: string
          format: date-time
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
        last_occurrence:
          "$ref": "#/components/schemas/NullableOccurrence"
    SchedulePage:
      type: object
      additionalProperties: false
      required:
      - items
      - next_cursor
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/Schedule"
        next_cursor:
          type: string
          nullable: true
    Settings:
      type: object
      additionalProperties: false
      required:
      - base_env_from
      - allowed_env_from
      - browser_auth
      properties:
        base_env_from:
          "$ref": "#/components/schemas/EnvFrom"
        allowed_env_from:
          "$ref": "#/components/schemas/EnvFrom"
        browser_auth:
          type: string
          enum:
          - api_only
          - anonymous
          - saml
    PreviewInput:
      type: object
      additionalProperties: false
      required:
      - cron
      - timezone
      properties:
        cron:
          type: string
          minLength: 1
        timezone:
          type: string
          minLength: 1
    Preview:
      type: object
      additionalProperties: false
      required:
      - times
      properties:
        times:
          type: array
          minItems: 5
          maxItems: 5
          items:
            type: string
            format: date-time
...
    Occurrence:
      type: object
      additionalProperties: false
      required:
      - id
      - schedule_id
      - scheduled_at
      - state
      - created_at
      - updated_at
      - session_id
      - run_id
      - run_status
      - observed_at
      - execution_started_at
      - finished_at
      - run_error_code
      - sync_error_code
      - error_code
      - attempts
      - next_attempt_at
      properties:
        id:
          type: string
          format: uuid
        schedule_id:
          type: string
          format: uuid
        scheduled_at:
          type: string
          format: date-time
        state:
          "$ref": "#/components/schemas/OccurrenceState"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        session_id:
          type: string
          nullable: true
          format: uuid
        run_id:
          type: string
          nullable: true
          format: uuid
        run_status:
          type: string
          nullable: true
        observed_at:
          type: string
          nullable: true
          format: date-time
        execution_started_at:
          type: string
          nullable: true
          format: date-time
        finished_at:
          type: string
          nullable: true
          format: date-time
        run_error_code:
          type: string
          nullable: true
        sync_error_code:
          type: string
          nullable: true
        error_code:
          type: string
          nullable: true
        attempts:
          type: integer
          minimum: 0
        next_attempt_at:
          type: string
          nullable: true
          format: date-time
    OccurrencePage:
      type: object
      additionalProperties: false
      required:
      - items
      - next_cursor
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/Occurrence"
        next_cursor:
          type: string
          nullable: true
    OccurrenceResult:
      type: object
      additionalProperties: false
      required:
      - run_status
      - fetched_at
      - error
      - final_message
      properties:
        run_status:
          type: string
        fetched_at:
          type: string
          format: date-time
        error:
          type: object
          nullable: true
          additionalProperties: false
          required:
          - code
          - message
          - phase
          properties:
            code:
              type: string
            message:
              type: string
            phase:
              type: string
              nullable: true
        final_message:
          type: object
          nullable: true
          additionalProperties: false
          required:
          - id
          - text
          - created_at
          properties:
            id:
              type: string
              format: uuid
            text:
              type: string
            created_at:
              type: string
              format: date-time
    OccurrenceState:
      type: string
      enum:
      - pending
      - dispatching
      - accepted
      - skipped
      - failed
      - cancelled
    NullableOccurrence:
      type: object
      additionalProperties: false
      required:
      - id
      - schedule_id
      - scheduled_at
      - state
      - created_at
      - updated_at
      - session_id
      - run_id
      - run_status
      - observed_at
      - execution_started_at
      - finished_at
      - run_error_code
      - sync_error_code
      - error_code
      - attempts
      - next_attempt_at
      properties:
        id:
          type: string
          format: uuid
        schedule_id:
          type: string
          format: uuid
        scheduled_at:
          type: string
          format: date-time
        state:
          "$ref": "#/components/schemas/OccurrenceState"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        session_id:
          type: string
          nullable: true
          format: uuid
        run_id:
          type: string
          nullable: true
          format: uuid
        run_status:
          type: string
          nullable: true
        observed_at:
          type: string
          nullable: true
          format: date-time
        execution_started_at:
          type: string
          nullable: true
          format: date-time
        finished_at:
          type: string
          nullable: true
          format: date-time
        run_error_code:
          type: string
          nullable: true
        sync_error_code:
          type: string
          nullable: true
        error_code:
          type: string
          nullable: true
        attempts:
          type: integer
          minimum: 0
        next_attempt_at:
          type: string
          nullable: true
          format: date-time
      nullable: true